requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pytest
django-pytest